# openapi.json
{
  "components": {
    "schemas": {
...
      "BoundingBox": {
        "description": "Object detection bounding box",
        "properties": {
          "height": {
            "title": "Height",
            "type": "number"
          },
          "width": {
            "title": "Width",
            "type": "number"
          },
          "x": {
            "title": "X",
            "type": "number"
          },
          "y": {
            "title": "Y",
            "type": "number"
          }
        },
        "required": [
          "x",
          "y",
          "width",
          "height"
        ],
        "title": "BoundingBox",
        "type": "object"
      },
...
      "DetectedObject": {
        "description": "Detected object with classification",
        "properties": {
          "angular_width_degrees": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 179.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Detector box width projected through the configured camera FOV.",
            "title": "Angular Width Degrees"
          },
          "bounding_box": {
            "$ref": "#/components/schemas/BoundingBox"
          },
          "class_name": {
            "title": "Class Name",
            "type": "string"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "distance_estimate": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 100.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance Estimate"
          },
          "object_id": {
            "title": "Object Id",
            "type": "string"
          },
          "relative_bearing": {
            "anyOf": [
              {
                "maximum": 180.0,
                "minimum": -180.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relative Bearing"
          },
          "semantic_cost_multiplier": {
            "default": 1.0,
            "description": "Route-cost inflation only; never reduces geometric or ToF safety.",
            "minimum": 1.0,
            "title": "Semantic Cost Multiplier",
            "type": "number"
          },
          "tracking_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tracking Id"
          }
        },
        "required": [
          "object_id",
          "class_name",
          "confidence",
          "bounding_box"
        ],
        "title": "DetectedObject",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InferenceResult": {
        "description": "Results from AI model inference",
        "properties": {
          "classification_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Classification Confidence"
          },
          "classification_result": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Classification Result"
          },
          "confidence_threshold": {
            "default": 0.5,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence Threshold",
            "type": "number"
          },
          "detected_objects": {
            "items": {
              "$ref": "#/components/schemas/DetectedObject"
            },
            "title": "Detected Objects",
            "type": "array"
          },
          "inference_id": {
            "minLength": 1,
            "title": "Inference Id",
            "type": "string"
          },
          "inference_time_ms": {
            "default": 0.0,
            "minimum": 0.0,
            "title": "Inference Time Ms",
            "type": "number"
          },
          "input_frame_id": {
            "minLength": 1,
            "title": "Input Frame Id",
            "type": "string"
          },
          "input_height": {
            "exclusiveMinimum": 0.0,
            "title": "Input Height",
            "type": "integer"
          },
          "input_width": {
            "exclusiveMinimum": 0.0,
            "title": "Input Width",
            "type": "integer"
          },
          "model_name": {
            "minLength": 1,
            "title": "Model Name",
            "type": "string"
          },
          "model_runtime": {
            "default": "unknown",
            "minLength": 1,
            "title": "Model Runtime",
            "type": "string"
          },
          "model_sha256": {
            "anyOf": [
              {
                "pattern": "^[0-9a-f]{64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Sha256"
          },
          "model_version": {
            "default": "1.0",
            "title": "Model Version",
            "type": "string"
          },
          "postprocessing_time_ms": {
            "default": 0.0,
            "minimum": 0.0,
            "title": "Postprocessing Time Ms",
            "type": "number"
          },
          "preprocessing_time_ms": {
            "default": 0.0,
            "minimum": 0.0,
            "title": "Preprocessing Time Ms",
            "type": "number"
          },
          "segmentation_mask": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Segmentation Mask"
          },
          "source_frame_timestamp": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Frame Timestamp"
          },
          "task": {
            "$ref": "#/components/schemas/InferenceTask"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          },
          "total_time_ms": {
            "default": 0.0,
            "minimum": 0.0,
            "title": "Total Time Ms",
            "type": "number"
          }
        },
        "required": [
          "inference_id",
          "task",
          "model_name",
          "input_frame_id",
          "input_width",
          "input_height"
        ],
        "title": "InferenceResult",
        "type": "object"
      },
      "InferenceTask": {
        "description": "Types of AI inference tasks",
        "enum": [
          "obstacle_detection",
          "grass_quality",
          "boundary_detection",
          "navigation_aid",
          "object_classification",
          "safety_monitoring"
        ],
        "title": "InferenceTask",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Autonomous robotic lawn mower backend API",
    "title": "LawnBerry Pi v2",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v2/ai/inference": {
      "post": {
        "description": "Run uploaded-image inference in the embedded SIM/CI runtime.\n\nHardware mode keeps inference in the standalone camera owner, so this\ndiagnostic route returns 503 instead of starting a competing backend\ninference path. Read `/api/v2/ai/perception/latest` for hardware results.",
        "operationId": "run_uploaded_inference_api_v2_ai_inference_post",
        "parameters": [
          {
            "in": "query",
            "name": "task",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/InferenceTask",
              "default": "obstacle_detection"
            }
          },
          {
            "in": "query",
            "name": "confidence_threshold",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 1.0,
                  "minimum": 0.0,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Confidence Threshold"
            }
          },
          {
            "in": "query",
            "name": "frame_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Frame Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InferenceResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Uploaded Inference"
      }
    },
    "/api/v2/ai/inference/latest": {
      "post": {
        "description": "Run latest-frame inference in the embedded SIM/CI runtime.\n\nHardware mode performs automatic sampled inference in the standalone\ncamera owner, so this diagnostic route returns 503. Read\n`/api/v2/ai/perception/latest` for the latest hardware result.",
        "operationId": "run_latest_frame_inference_api_v2_ai_inference_latest_post",
        "parameters": [
          {
            "in": "query",
            "name": "task",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/InferenceTask",
              "default": "obstacle_detection"
            }
          },
          {
            "in": "query",
            "name": "confidence_threshold",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 1.0,
                  "minimum": 0.0,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Confidence Threshold"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InferenceResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Latest Frame Inference"
      }
    },
...
    "/api/v2/ai/results/recent": {
      "get": {
        "description": "Return recent inference results, newest first.",
        "operationId": "get_recent_results_api_v2_ai_results_recent_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InferenceResult"
                  },
                  "title": "Response Get Recent Results Api V2 Ai Results Recent Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Recent Results"
      }
    },
...
  }
}